Remove theme from utils and visual tests
