+++ packages/components/src/OperationalUI/OperationalUI.tsx
@@ -2,8 +2,8 @@ import * as React from "react"
 import { injectGlobal } from "emotion"
 import { ThemeProvider } from "emotion-theming"
 
-import constants from "../utils/constants"
-import { baseStylesheet } from "@operational/utils"
+import constants, { OperationalStyleConstants } from "../utils/constants"
+import { darken } from "@operational/utils"
 
 export interface Props {
   /** Children */
@@ -18,9 +18,41 @@ export interface Props {
 
 const { Provider, Consumer } = React.createContext({})
 
+export const baseStylesheet = (theme: OperationalStyleConstants): string => `
+* {
+  box-sizing: border-box;
+  text-rendering: optimizeLegibility;
+  -webkit-font-smoothing: antialiased;
+  font-smoothing: antialiased;
+}
+
+html,
+body {
+  margin: 0;
+  padding: 0;
+  font-family: ${theme.deprecated.fontFamily};
+  font-size: 13px;
+  height: 100%;
+}
+
+body {
+  background-color: ${theme.deprecated.colors.background};
+}
+
+a:link,
+a:visited {
+  color: ${theme.deprecated.colors.info};
+  text-decoration: none;
+}
+
+a:hover: {
+  color: ${darken(theme.deprecated.colors.info, 5)};
+}
+`
+
 class OperationalUI extends React.Component<Props, {}> {
   componentDidMount() {
-    this.props.withBaseStyles && injectGlobal(baseStylesheet(constants.deprecated))
+    this.props.withBaseStyles && injectGlobal(baseStylesheet(constants))
   }
 
   render() {
