+++ packages/components/src/Select/Select.tsx
@@ -4,8 +4,8 @@ import styled from "react-emotion"
 import { Label, LabelText, inputFocus } from "../utils/mixins"
 import SelectOption from "./Select.Option"
 import SelectFilter from "./Select.Filter"
-import { readableTextColor, fadeIn, resetTransform, expandColor } from "@operational/utils"
-import { OperationalStyleConstants } from "../utils/constants"
+import { readableTextColor, fadeIn, resetTransform } from "@operational/utils"
+import { OperationalStyleConstants, expandColor } from "../utils/constants"
 
 export type Value = number | string
