+++ packages/components/src/SidenavHeader/SidenavHeader.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import styled, { Interpolation } from "react-emotion"
-import { OperationalStyleConstants } from "../utils/constants"
-import { expandColor, fadeIn } from "@operational/utils"
+import { OperationalStyleConstants, expandColor } from "../utils/constants"
+import { fadeIn } from "@operational/utils"
 import { isModifiedEvent } from "../utils"
 import { Icon, IconName, ContextConsumer, Context } from "../"
 import { Props as SidenavItemProps } from "../SidenavItem/SidenavItem"
