+++ packages/components/src/utils/constants.ts
@@ -172,8 +172,33 @@ const constants = {
   deprecated: operational,
 }
 
+/*
+ * Expands a color expressed either as a custom hex value
+ * or a color key to pick from within the style constants object.
+*/
+export const expandColor = (
+  theme: OperationalStyleConstants,
+  color?: keyof OperationalStyleConstants["color"] | string,
+): string | null => {
+  if (!color) {
+    return null
+  }
+  const hexRegEx = /(^#[0-9A-F]{6}$)|(^#[0-9A-F]{3}$)|currentColor/i
+  const isHex = hexRegEx.test(color)
+  if (isHex) {
+    return color
+  }
+
+  /**
+   * This function is typically used in checks.
+   * If falsy, it returns a fallback color, hence
+   * the empty string return for a falsy value.
+   */
+  return (theme.color as any)[color] || ""
+}
+
 export type OperationalStyleConstants = Readonly<typeof constants>
 
 export default constants
 
-export { expandColor } from "./constants/deprecatedTheme"
+export { expandColor as deprecatedExpandColor } from "./constants/deprecatedTheme"
