+++ packages/visual-tests/src/MarathonRenderer.tsx
@@ -1,8 +1,7 @@
 import * as React from "react"
 import styled from "react-emotion"
 
-import { Icon } from "@operational/components"
-import { OperationalStyleConstants, operational } from "@operational/theme"
+import { Icon, OperationalStyleConstants } from "@operational/components"
 import { MarathonRenderer } from "./Marathon"
 
 const Container = styled("ul")({
@@ -59,9 +58,9 @@ const MarathonRendererComponent = ({ results, ref }: MarathonRenderer) => (
       {results.map((result, index) => {
         const content = result.isCompleted ? (
           result.errors.length > 0 ? (
-            <Icon name="X" size={12} color={operational.colors.error} />
+            <Icon name="X" size={12} color="#F00" />
           ) : (
-            <Icon name="Check" size={12} color={operational.colors.success} />
+            <Icon name="Check" size={12} color="#0F0" />
           )
         ) : (
           <Icon name="MoreHorizontal" size={12} />
